fix: keep nanobind's own warnings from breaking strict CI builds
nanobind ships a small compiled library (unlike header-only pybind11),
so its own source and the Python headers it pulls in became subject to
this project's strict, warnings-as-errors compiler flags once linked in.
nanobind's include directories are now marked as SYSTEM headers, which
covers warnings that originate in headers, and treating warnings as
errors is disabled specifically for nanobind's compiled library for the
handful of warnings that originate directly in its own source rather
than in a header.

Also fixes a real, unrelated sign-conversion warning and a missing
unused-parameter marker introduced by the pybind11 to nanobind switch.

+++ python/type_caster/tensor.hpp
@@ -345,7 +345,7 @@ private:
 
         for (std::size_t i = 0; i != N; ++i) {
             shape[i] =
-                static_cast<std::size_t>(t.dimension(static_cast<int>(i)));
+                static_cast<std::size_t>(t.dimension(i));
             elementStrides[i] = static_cast<std::int64_t>(
                 byteStrides[i] /
                 static_cast<Eigen::DenseIndex>(sizeof(TensorScalar)));
@@ -403,7 +403,7 @@ private:
 
         for (std::size_t i = 0; i != NumDimensions; ++i) {
             shape[i] =
-                static_cast<std::size_t>(p->dimension(static_cast<int>(i)));
+                static_cast<std::size_t>(p->dimension(i));
             elementStrides[i] = static_cast<std::int64_t>(
                 byteStrides[i] /
                 static_cast<Eigen::DenseIndex>(sizeof(Scalar)));
